fix(ui): update test import paths after moving WIP pages to _wip/
Test files for MotionCapture, PuttingGreen, and VideoAnalyzer still referenced
the old ./PageName import paths after the components were relocated to _wip/.
Updated all three to import from ./_wip/PageName, resolving TS2307 errors.
