+++ ui/src/pages/VideoAnalyzer.test.tsx
@@ -6,7 +6,7 @@
 
 import { describe, it, expect } from 'vitest';
 
-import type { VideoAnalysisResult, PoseFrame, TaskStatus } from './VideoAnalyzer';
+import type { VideoAnalysisResult, PoseFrame, TaskStatus } from './_wip/VideoAnalyzer';
 
 describe('VideoAnalyzer data structures', () => {
   it('should parse video analysis result', () => {
